care_gap_mcp/pyproject.toml
[project]
name = "care-gap-mcp"
version = "0.1.0"
description = "USPSTF-aligned care-gap finder + patient outreach drafter, exposed as an MCP server with SHARP/FHIR context."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp[apps]>=3.2.4",
    "fhir.resources>=8.2.0",
    "httpx>=0.28.1",
    "pydantic>=2.13.3",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
]
